update readme for new api
